Removido alguns controllers desnecessarios
